+++ productAndOrderService/pom.xml
@@ -24,9 +24,8 @@
         </dependency>
         <dependency>
             <groupId>org.example</groupId>
-            <artifactId>userService</artifactId>
+            <artifactId>userSDK</artifactId>
             <version>1.0-SNAPSHOT</version>
-            <scope>compile</scope>
         </dependency>
     </dependencies>
