+++ productAndOrderService/src/main/java/Managers/OrdersManager.java
@@ -4,38 +4,24 @@ import Entities.Order;
 
 import Handlers.MorphiaHandler;
 import com.example.dummy.UserExistClient;
-import com.example.dummy.UserExistServer;
 
 import java.util.List;
 
 public class OrdersManager {
 
+    UserExistClient userExistClient;
     MorphiaHandler morphiaHandler;
     ProductsManager productsManager;
-    UserExistClient userExistClient;
-    UserExistServer userExistServer;
 
     public OrdersManager() {
         morphiaHandler = new MorphiaHandler();
-        productsManager = new ProductsManager();
         userExistClient = new UserExistClient();
-        userExistServer = new UserExistServer();
+        productsManager = new ProductsManager();
     }
 
     public String insert(Order order) {
         if (order.getProducts().length == 0)
             return "No Products added";
-        // Starting gRPC Server
-        new Thread(new Runnable() {
-            public void run() {
-                try {
-                    userExistServer.server();
-                } catch (Exception e) {
-                    e.printStackTrace();
-                }
-            }
-        }).start();
-
         if (!userExistClient.doesUserExist(order.getUserId()))
             return "user doesn't exist";
         for (int x : order.getProducts()) {
