+++ userSDK/src/main/java/com/example/dummy/UserExistServiceImpl.java
@@ -13,7 +13,7 @@ public class UserExistServiceImpl extends UserExistServiceGrpc.UserExistServiceI
         UsersManager usersManager = new UsersManager();
         List<Object> lis = usersManager.getUserById(request.getUserId());
         boolean user = true;
-        if(lis.size()==0)
+        if (lis.size() == 0)
             user = false;
 
         UserExist.UserExistResponse response = UserExist.UserExistResponse.newBuilder()
